+++ Assets/VRM10/Runtime/Migration/MigrationException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace UniVRM10
+{
+    public class MigrationException : Exception
+    {
+        public MigrationException(string key, string value) : base($"{key}: {value}")
+        {
+        }
+    }
+}
