+++ Assets/VRM10/Runtime/Migration/MigrationException.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 082de46c993eb764eb020101c113824a
+guid: 09e748ecf102bb44a96f0b640a6a3ed1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
