+++ Assets/VRM10/Tests/MigrationTests.cs
@@ -46,14 +46,14 @@ namespace UniVRM10
             var vrm0Bytes = File.ReadAllBytes(AliciaPath);
             var vrm0Json = GetVRM0(vrm0Bytes);
 
-            var vrm1 = Migration.Migrate(vrm0Bytes);
+            var vrm1 = MigrationVrm.Migrate(vrm0Bytes);
             var glb = UniGLTF.Glb.Parse(vrm1);
             var json = glb.Json.Bytes.ParseAsJson();
             var gltf = UniGLTF.GltfDeserializer.Deserialize(json);
 
-            MigrationCheck.Check(vrm0Json, GetExtension(gltf.extensions, UniGLTF.Extensions.VRMC_vrm.VRMC_vrm.ExtensionNameUtf8,
+            MigrationVrm.Check(vrm0Json, GetExtension(gltf.extensions, UniGLTF.Extensions.VRMC_vrm.VRMC_vrm.ExtensionNameUtf8,
                 UniGLTF.Extensions.VRMC_vrm.GltfDeserializer.Deserialize));
-            MigrationCheck.Check(vrm0Json, GetExtension(gltf.extensions, UniGLTF.Extensions.VRMC_springBone.VRMC_springBone.ExtensionNameUtf8,
+            MigrationVrm.Check(vrm0Json, GetExtension(gltf.extensions, UniGLTF.Extensions.VRMC_springBone.VRMC_springBone.ExtensionNameUtf8,
                 UniGLTF.Extensions.VRMC_springBone.GltfDeserializer.Deserialize), gltf.nodes);
         }
     }
